Update to compute_antenna_azimuth_direction
More robust way of handling antenna direction and whether or not SquintImage is present. Uses dicts to parse the angular antenna directions rather than if statements. Now also allows for the Mid antenna to be computed.

Small change to add_antenna_baseline to return only the baseline variable rather than the whole dataset for issue #182