L1 and L2 lat/lon coordinates fix
Fixes the issue with NaNs appearing in the L2 dataset latitude and longitude coordinates, which prevented their use in plotting etc.

Addresses and fixes #193, as well as #185 and #181 now that joining the beams by 'outer' will not result in these NaNs creeping in.

Fixed by means of xarray's .dropna function and working along the CrossRange dimension. As long as the lack of LatImage / LonImage data is always in this dimension (and not in GroundRange, i.e. not with sloping ends) then this will work robustly with OSCAR data.