Compute NRCS
Function to compute Sigma0 using the new nscat4ds code in .gmfs.nrcs. Its a little stripped-down as I couldn't get the way we treat the antenna polarization in the wasv code to work here, however this does deal with the combined antenna dimension.

Also included is a small change to compute_radial_surface_current that allows the null mid antenna data to pass through, however it still needs to be adapted to accept combined beam datasets.

Finally an example notebook is included.