Small change to compute_wasv
The commented-out line was preventing compute_wasv to function in multiprocessing mode due to the fact that L1.Polarization.size == 1 for an MPI pixel. Thus ind was not populated and crashed further down at line 198 with `Key Error: 'VV'`. This small change allows the code to run with yurovsky but I have left the line commented in as a reminder