Update level1.py
Added a new function to apply the phase sign convention to the interferograms

Changes to L1B processor to apply the phase sign convention

Small changes to L1B processor as bug-fixes. Line 776 to 782 threw errors as ds_ml is a dict and not a DataSet. Moved these operations to after merge_beams to work on the new L1B dataset